refactor(suno): 例外 import を canonical owner へ移行する (#3959)
